requirements.txt
langchain-community
langchain-groq
cassio
PyPDF2
sentence-transformers
gradio
python-dotenv
datasets
